Extract atoms
