+++ packages/react/src/Components/Atoms/index.js
@@ -1,0 +1,6 @@
+export { default as Button } from './Button';
+export { default as CanvasImage } from './CanvasImage';
+export { default as CanvasStrockeRect } from './CanvasStrokeRect';
+export { default as Image } from './Image';
+export { default as TableRow } from './TableRow';
+export { default as Video } from './Video';
