+++ packages/react/src/Components/InfoTable.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 
 // Local imports
-import TableRow from './TableRow';
+import { TableRow } from './Atoms';
 import { Row } from './Layout';
 
 const InfoTable = () => {
